requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==4.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
et_xmlfile==2.0.0
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.48.2
h11==0.14.0
idna==3.10
mypy==1.13.0
mypy-extensions==1.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
rsa==4.9
selenium==4.26.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
types-beautifulsoup4==4.12.0.20241020
types-html5lib==1.1.11.20241018
types-openpyxl==3.1.5.20241025
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
